Remove circular dependency between rules and atrules parsers
