+++ src/index.ts
@@ -1,8 +1,10 @@
 import { Root, Plugin } from 'postcss';
-import { PluginOptions } from '@types';
+import { PluginOptions, Parsers } from '@types';
 import { initStore } from '@data/store';
-import { parseKeyFrames, parseAtRules } from '@parsers/atrules';
+import { parseAtRules } from '@parsers/atrules';
 import { parseRules } from '@parsers/rules';
+import { parseKeyFrames } from '@parsers/keyframes';
+import { parseDeclarations } from '@parsers/declarations';
 import {
     appendRules,
     appendKeyFrames,
@@ -10,14 +12,21 @@ import {
 } from '@utilities/rules';
 import { clean } from '@utilities/clean';
 
+const parsers: Parsers = {
+    parseRules,
+    parseAtRules,
+    parseKeyFrames,
+    parseDeclarations
+};
+
 function postcssRTLCSS (options: PluginOptions = {}): Plugin {
     return ({
         postcssPlugin: 'postcss-rtlcss',
         Once(css: Root): void {
             initStore(options);
             parseKeyFrames(css);
-            parseAtRules(css);
-            parseRules(css);
+            parseAtRules(parsers, css);
+            parseRules(parsers, css);
             parseRuleNames();
             appendRules();
             appendKeyFrames();
